migeate(W-14169209): spaces: Upgrade spaces:transfer (#2837)
* migeate(W-14169209): spaces: Upgrade spaces:transfer

* Updated help and unit tests
